back-end/mealplan_api/pyproject.toml
[tool.poetry]
name = "mealplan_api"
version = "1.1.0"
description = "siksa_clock graphql mvp"
authors = ["Hyeonproject <axzswq@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
strawberry-graphql = {extras = ["fastapi"], version = "^0.124.0"}
none = "^0.1.0-alpha.2"
httpx = "^0.23.0"
pytest = "^7.1.2"
anyio = "^3.6.1"
requests = "^2.28.0"
password = "^0.2"
schema = "^0.7.5"
crud = "^0.1"
pyright = "^1.1.276"
black = "^22.10.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.2"}
psycopg2 = "^2.9.5"

[tool.poetry.dev-dependencies]


[tool.black]
line-length = 120
target-version = ['py3.11']
include = '\.pyi?$'



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
